Included error messages
The case where only a blank screen shows when first opening app with no
network connection seemed not to be a problem. However, now there is a
message if the stocks may be out of date. Also, the app no longer
crashes with an invalid stock input. Instead, it creates a new stock
label stating that it's invalid.

Should fix Jamal's issue and Xaio-lu's issue